Trivial, remove junk char

+++ app.py
@@ -66,7 +66,7 @@ def index():
             return render_template('index.html', output_text=errMessage)
 
         jsonSize = len(json.dumps(dataJSON,  separators=(',', ':')))
-        hexSize  = len(hexText)/2
+        hexSize  = int(len(hexText)/2)
         compressedPercent = round(100 - (hexSize/jsonSize)*100, 2)
         return render_template('index.html', output_text=outputText, hex_text=hexText, json_length=jsonSize, hex_length=hexSize, 
                                compressed_percent = compressedPercent, sid_text_raw=sidTextRaw, data_text_raw=dataTextRaw)
